Move starting point one layer up to AbstractMapRenderer
