+++ src/main/java/com/github/johnnyjayjay/spigotmaps/rendering/ImageRenderer.java
@@ -20,7 +20,6 @@ import java.util.function.Predicate;
 public class ImageRenderer extends AbstractMapRenderer {
 
     private BufferedImage image;
-    private Point startingPoint;
 
     private ImageRenderer(
             Set<Player> receivers,
@@ -29,9 +28,8 @@ public class ImageRenderer extends AbstractMapRenderer {
             BufferedImage image,
             Point startingPoint
     ) {
-        super(receivers, renderOnce, precondition);
+        super(startingPoint, receivers, renderOnce, precondition);
         this.image = image;
-        this.startingPoint = startingPoint;
     }
 
     @Override
@@ -55,24 +53,6 @@ public class ImageRenderer extends AbstractMapRenderer {
     public void setImage(BufferedImage image) {
         Checks.checkNotNull(image, "Image");
         this.image = ImageTools.copyOf(image);
-    }
-
-    /**
-     * Returns a copy of the point where the renderer begins to render text on a map.
-     */
-    public Point getStartingPoint() {
-        return new Point(startingPoint);
-    }
-
-    /**
-     * Sets the point on the map where this renderer should start rendering.
-     *
-     * @param startingPoint the point to set.
-     * @throws IllegalArgumentException if the given point cannot be applied to a minecraft map or is null.
-     */
-    public void setStartingPoint(Point startingPoint) {
-        Checks.checkStartingPoint(startingPoint);
-        this.startingPoint = new Point(startingPoint);
     }
 
     /**
@@ -114,7 +94,6 @@ public class ImageRenderer extends AbstractMapRenderer {
     public static class Builder extends AbstractMapRenderer.Builder<ImageRenderer, Builder> {
 
         private BufferedImage image = null;
-        private Point startingPoint = new Point();
 
         private Builder() {
         }
@@ -135,7 +114,6 @@ public class ImageRenderer extends AbstractMapRenderer {
         @Override
         public ImageRenderer build() {
             super.check();
-            Checks.checkStartingPoint(startingPoint);
             Checks.checkNotNull(image, "Image");
             return new ImageRenderer(receivers, precondition, renderOnce, image, startingPoint);
         }
@@ -154,23 +132,6 @@ public class ImageRenderer extends AbstractMapRenderer {
          */
         public Builder image(BufferedImage image) {
             this.image = ImageTools.copyOf(image);
-            return this;
-        }
-
-        /**
-         * Sets the coordinates (via a {@link Point} determining where to begin drawing the image on the map.
-         * <p>
-         * This makes a defensive copy of the {@link Point}, so changes to the argument will not have any
-         * effect on this instance.
-         * <p>
-         * This is not required. By default, it will start drawing from the upper left corner (0, 0).
-         *
-         * @param point a non-{@code null} {@link Point} representing the coordinates, i.e. where to begin drawing.
-         * @return this.
-         * @see ImageTools#MINECRAFT_MAP_SIZE
-         */
-        public Builder startingPoint(Point point) {
-            this.startingPoint = new Point(point);
             return this;
         }
     }
